# swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    GenericError:
        properties:
            message:
                type: string
                x-go-name: Message
        type: object
        x-go-package: github.com/product-api-microservice/handlers
...
info:
    description: Documentation for Product API
    title: of Product API
    version: 1.0.0
paths:
    /product/{id}:
        delete:
            description: Delete Product for given ID
            operationId: deleteProduct
            parameters:
                - description: The id of product to delete from DB
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    $ref: '#/responses/noContent'
                "404":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
...
produces:
    - application/json
responses:
    errorResponse:
        description: ""
        schema:
            $ref: '#/definitions/GenericError'
    noContent:
        description: ""
...
schemes:
    - http
swagger: "2.0"